refactor(skeleton): reorganize skeleton components into modular structure
- Split monolithic SkeletonLoader.jsx into organized subdirectories
- Create Utility/, Cards/, Pages/, and Classroom/ component groups
- Extract centralized SKELETON_THEMES for reusable theming
- Add index files for clean imports and better maintainability
- Update import paths to use new modular structure
- Preserve all existing functionality and SCSS-based styling
