+++ frontend/src/components/UI/Skeleton/Cards/index.js
@@ -1,0 +1,5 @@
+export { CardSkeleton } from './CardSkeleton';
+export { WordCardSkeleton } from './WordCardSkeleton';
+export { ClassroomCardSkeleton } from './ClassroomCardSkeleton';
+export { VocabularyListCardSkeleton } from './VocabularyListCardSkeleton';
+export { ViewClassroomCardSkeleton } from './ViewClassroomCardSkeleton';
