+++ frontend/src/components/UI/Skeleton/Classroom/index.js
@@ -1,0 +1,5 @@
+export { ViewClassroomSkeleton } from './ViewClassroomSkeleton';
+export { LearnersListSkeleton } from './LearnersListSkeleton';
+export { PendingRequestSkeleton } from './PendingRequestSkeleton';
+export { AssignmentPageSkeleton } from './AssignmentPageSkeleton';
+export { AssignmentDetailSkeleton } from './AssignmentDetailSkeleton';
