+++ frontend/src/components/UI/Skeleton/Pages/index.js
@@ -1,0 +1,7 @@
+export { CarouselSectionSkeleton } from './CarouselSectionSkeleton';
+export { SearchResultsGridSkeleton } from './SearchResultsGridSkeleton';
+export { WordFormSkeleton } from './WordFormSkeleton';
+export { ListFormSkeleton } from './ListFormSkeleton';
+export { ViewListSkeleton, WordListSkeleton } from './ViewListSkeleton';
+export { ProfileSkeleton } from './ProfileSkeleton';
+export { OverviewListSkeleton } from './OverviewListSkeleton';
