+++ frontend/src/components/UI/Skeleton/Utility/index.js
@@ -1,0 +1,4 @@
+export { ContentSkeleton } from './ContentSkeleton';
+export { AvatarSkeleton } from './AvatarSkeleton';
+export { ButtonSkeleton } from './ButtonSkeleton';
+export { TableRowSkeleton } from './TableRowSkeleton';
