+++ frontend/src/components/UI/Skeleton/index.js
@@ -1,0 +1,14 @@
+// Themes
+export { SKELETON_THEMES } from './themes';
+
+// Utility Components
+export * from './Utility';
+
+// Card Components  
+export * from './Cards';
+
+// Page Components
+export * from './Pages';
+
+// Classroom Components
+export * from './Classroom';
