+++ frontend/src/components/UI/Skeleton/themes.js
@@ -1,0 +1,26 @@
+// ====================================================================
+// SKELETON THEMES
+// ====================================================================
+
+export const SKELETON_THEMES = {
+  default: {
+    baseColor: "#f3f4f6",
+    highlightColor: "#e5e7eb"
+  },
+  classroom: {
+    baseColor: "#f1ede4", 
+    highlightColor: "#e8ddc7"
+  },
+  cream: {
+    baseColor: "#FFFCF7",
+    highlightColor: "#F7F1E3"
+  },
+  vocabCard: {
+    baseColor: "#F9F3E4",
+    highlightColor: "#F1EBD8"
+  },
+  overview: {
+    baseColor: "#fff6e9",
+    highlightColor: "#ffe8d1"
+  }
+};
