+++ frontend/src/components/index.jsx
@@ -18,7 +18,7 @@ export { default as AssignSubMenu } from "./Navigation/AssignSubMenu.jsx";
 // UI Components
 export { default as ConfirmModal } from "./UI/ConfirmModal.jsx";
 export { default as LoadingCursor } from "./UI/LoadingCursor.jsx";
-export * from "./UI/SkeletonLoader.jsx";
+export * from "./UI/Skeleton";
 
 // Providers
 export { useConfirm, ConfirmProvider } from "./Providers/ConfirmProvider.jsx";
